use dashkit as default theme for backend and edit default resources views

+++ db/migrate/20220125092685_drop_spree_store_credit_update_reasons.spree.rb
@@ -1,4 +1,5 @@
 # frozen_string_literal: true
+
 # This migration comes from spree (originally 20190220093635)
 
 class DropSpreeStoreCreditUpdateReasons < ActiveRecord::Migration[5.1]
@@ -9,9 +10,7 @@ class DropSpreeStoreCreditUpdateReasons < ActiveRecord::Migration[5.1]
   # of migration 20180710170104 would fail with `table already exists` , as
   # it handles itself the add/remove of this table and column.
   def up
-    if table_exists? :spree_store_credit_update_reasons
-      drop_table :spree_store_credit_update_reasons
-    end
+    drop_table :spree_store_credit_update_reasons if table_exists? :spree_store_credit_update_reasons
 
     if column_exists? :spree_store_credit_events, :update_reason_id
       remove_column :spree_store_credit_events, :update_reason_id
